requirements.txt
agent-framework
azure-ai-projects
azure-identity
python-dotenv
rich
